requirements.txt
flask==1.1.2
Flask-SQLAlchemy==2.4.3
flask-migrate==2.5.3
python-dotenv==0.13.0
flask-jwt-extended==0.3.2
flask-restful==0.3.8
marshmallow==3.6.1
flask-marshmallow==0.13.0
marshmallow-sqlalchemy==0.23.1